pyproject.toml
[tool.poetry]
package-mode = false
name = "wamex-rag-assistant"

[project]
name = "wamex-rag-assistant"
version = "0.1.0"
description = "True serverless RAG architecture using Amazon Bedrock, S3, and an ephemeral FAISS vector store to extract natural language insights from WAMEX geological reports."
authors = [
    {name = "Evan G",email = "evan.f.l.gloria@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "boto3 (>=1.42.61,<2.0.0)",
    "langchain (<0.4.0)",
    "langchain-community (<0.4.0)",
    "langchain-aws (<0.3.0)",
    "pypdf (>=6.7.5,<7.0.0)",
    "numpy (<2.0.0)",
    "faiss-cpu (==1.7.4)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[dependency-groups]
frontend = [
    "streamlit (>=1.55.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]
